// libs/core/src/lib/hass/generated/zeroconf.ts
/**
 * Automatically generated by hassfest.

To update, run python3 -m script.hassfest

 */

// fmt: off

export const ZEROCONF = {
    "_Volumio._tcp.local.": [
        {
            "domain": "volumio"
        }
    ],
    "_api._udp.local.": [
        {
            "domain": "guardian"
        }
    ],
    "_axis-video._tcp.local.": [
        {
            "domain": "axis",
            "macaddress": "00408C*"
        },
        {
            "domain": "axis",
            "macaddress": "ACCC8E*"
        },
        {
            "domain": "axis",
            "macaddress": "B8A44F*"
        },
        {
            "domain": "doorbird",
            "macaddress": "1CCAE3*"
        }
    ],
    "_bond._tcp.local.": [
        {
            "domain": "bond"
        }
    ],
    "_daap._tcp.local.": [
        {
            "domain": "forked_daapd"
        }
    ],
    "_dkapi._tcp.local.": [
        {
            "domain": "daikin"
        }
    ],
    "_elg._tcp.local.": [
        {
            "domain": "elgato"
        }
    ],
    "_esphomelib._tcp.local.": [
        {
            "domain": "esphome"
        }
    ],
    "_googlecast._tcp.local.": [
        {
            "domain": "cast"
        }
    ],
    "_hap._tcp.local.": [
        {
            "domain": "homekit_controller"
        }
    ],
    "_homekit._tcp.local.": [
        {
            "domain": "homekit"
        }
    ],
    "_http._tcp.local.": [
        {
            "domain": "shelly",
            "name": "shelly*"
        }
    ],
    "_ipp._tcp.local.": [
        {
            "domain": "ipp"
        }
    ],
    "_ipps._tcp.local.": [
        {
            "domain": "ipp"
        }
    ],
    "_mediaremotetv._tcp.local.": [
        {
            "domain": "apple_tv"
        }
    ],
    "_miio._udp.local.": [
        {
            "domain": "xiaomi_aqara"
        },
        {
            "domain": "xiaomi_miio"
        }
    ],
    "_nut._tcp.local.": [
        {
            "domain": "nut"
        }
    ],
    "_plugwise._tcp.local.": [
        {
            "domain": "plugwise"
        }
    ],
    "_printer._tcp.local.": [
        {
            "domain": "brother",
            "name": "brother*"
        }
    ],
    "_spotify-connect._tcp.local.": [
        {
            "domain": "spotify"
        }
    ],
    "_ssh._tcp.local.": [
        {
            "domain": "smappee",
            "name": "smappee1*"
        },
        {
            "domain": "smappee",
            "name": "smappee2*"
        }
    ],
    "_touch-able._tcp.local.": [
        {
            "domain": "apple_tv"
        }
    ],
    "_viziocast._tcp.local.": [
        {
            "domain": "vizio"
        }
    ],
    "_wled._tcp.local.": [
        {
            "domain": "wled"
        }
    ],
    "_xbmc-jsonrpc-h._tcp.local.": [
        {
            "domain": "kodi"
        }
    ]
}

export const HOMEKIT = {
    "3810X": "roku",
    "4660X": "roku",
    "7820X": "roku",
    "819LMB": "myq",
    "AC02": "tado",
    "Abode": "abode",
    "BSB002": "hue",
    "C105X": "roku",
    "C135X": "roku",
    "Healty Home Coach": "netatmo",
    "Iota": "abode",
    "LIFX": "lifx",
    "Netatmo Relay": "netatmo",
    "PowerView": "hunterdouglas_powerview",
    "Presence": "netatmo",
    "Rachio": "rachio",
    "Socket": "wemo",
    "TRADFRI": "tradfri",
    "Welcome": "netatmo",
    "Wemo": "wemo",
    "iSmartGate": "gogogate2",
    "tado": "tado"
}