// libs/core/src/lib/hass/generated/dhcp.ts
/**
 * Automatically generated by hassfest.

To update, run python3 -m script.hassfest

 */

// fmt: off

export const DHCP = [
    {
        "domain": "august",
        "hostname": "connect",
        "macaddress": "D86162*"
    },
    {
        "domain": "august",
        "hostname": "connect",
        "macaddress": "B8B7F1*"
    },
    {
        "domain": "flume",
        "hostname": "flume-gw-*",
        "macaddress": "ECFABC*"
    },
    {
        "domain": "flume",
        "hostname": "flume-gw-*",
        "macaddress": "B4E62D*"
    },
    {
        "domain": "nest",
        "macaddress": "18B430*"
    },
    {
        "domain": "nexia",
        "hostname": "xl857-*",
        "macaddress": "000231*"
    },
    {
        "domain": "nuheat",
        "hostname": "nuheat",
        "macaddress": "002338*"
    },
    {
        "domain": "powerwall",
        "hostname": "1118431-*",
        "macaddress": "88DA1A*"
    },
    {
        "domain": "powerwall",
        "hostname": "1118431-*",
        "macaddress": "000145*"
    },
    {
        "domain": "rachio",
        "hostname": "rachio-*",
        "macaddress": "009D6B*"
    },
    {
        "domain": "rachio",
        "hostname": "rachio-*",
        "macaddress": "F0038C*"
    },
    {
        "domain": "rachio",
        "hostname": "rachio-*",
        "macaddress": "74C63B*"
    },
    {
        "domain": "ring",
        "hostname": "ring*",
        "macaddress": "0CAE7D*"
    },
    {
        "domain": "roomba",
        "hostname": "irobot-*",
        "macaddress": "501479*"
    },
    {
        "domain": "sense",
        "hostname": "sense-*",
        "macaddress": "009D6B*"
    },
    {
        "domain": "sense",
        "hostname": "sense-*",
        "macaddress": "DCEFCA*"
    },
    {
        "domain": "solaredge",
        "hostname": "target",
        "macaddress": "002702*"
    },
    {
        "domain": "somfy",
        "hostname": "gateway-*",
        "macaddress": "F8811A*"
    },
    {
        "domain": "somfy_mylink",
        "hostname": "somfy_*",
        "macaddress": "B8B7F1*"
    },
    {
        "domain": "tesla",
        "hostname": "tesla_*",
        "macaddress": "4CFCAA*"
    },
    {
        "domain": "tesla",
        "hostname": "tesla_*",
        "macaddress": "044EAF*"
    },
    {
        "domain": "tesla",
        "hostname": "tesla_*",
        "macaddress": "98ED5C*"
    }
]